Moved code into more cohesive modules
